==> R_scripts/05_rss_by_distance_calibration.R <==
###################################################################################
## Diane Klement
## January 31 2024
##
## Code to examine the relationship between RSS values and distance for a given study area using an exponential decay model
##
##
###################################################################################
###########################################################################################################################################################
##
##  Kristina L Paxton
##
##  April 1 2021
##
##    Code to examine the relationship between RSS values and distance for a given study area using an exponential decay model
##      -- Analysis in Ecology and Evolution Paper is based on data from a test with of a radio transmitter at 135 known locations distributed throughout the Guam Network 
##         -- At each test location a transmitter was held stationary for a 5-minute time period 
##            -- Removed first and last minute of each test to ensure that times matched between tests and the node network
##            -- For each test, calculated an average RSS value for the 3-min middle time period individually for each node that detected the transmitter
##        
##
##    1st step: Data preparation - isolates raw RSS data from node network that is associated with test data 
##      -- Creates the data file that was published with this paper at: https://doi.org/10.5066/P94LQWIE 
##            -- Data associated with this test is coded as 'A' in the column 'DataSet'
##    2nd step: Exponential Decay Model - uses the dataset created in step 1 to examine the relationship between RSS values and distance for a node network
##
##
##    Files Needed
##        1. TestInfo.csv: file with test information that is saved in the working directory defined below
##            - For each unique test location there should a row of information associated with each of the 3 inner minutes of that test
##                - meaning that each test will have 3 rows of information with each row identifying one of the 3 inner minutes of the test (see TestInfo_Example.csv)
##            - For this code, Diane's csv is: "TestInfo_Example.csv"
##            - Columns
##                -- TagId: unique identifier of the transmitter used during the specified test
##                -- TestId: unique identifier given to each of the unique locations where a test was conducted
##                -- Date: date when the specified test was conducted
##                -- Start.Time: time when the specified test was started
##                -- End.Time: time when the specified test was ended
##                -- Min: 1-minute time period of the specified test
##                -- Hour: hour of the the specified minute of the test
##                -- TestUTMx: Easting location of the specified test 
##                -- TestUTMy: Northing location of the specified test 
##
##       2. BeepData.rds: file with the raw RSS values collected by the node network during the time period of the test. The file should be saved in the working directory defined below
##            -- output file from Import_beep.data.Github.R script (see BeepData_Example.rds)
##            -- If using a file created from another source the following columns are needed in the specified format: 
##                -- TagId: Factor identifying the unique code of a tag
##                -- Time.local: POSIXct value in the format: "2021-12-29 09:02:51" that identifies the unique datetime stamp (IN THE LOCAL TIME ZONE OF THE STUDY) when the tag was detected by the specified node
##                    ***** If you don't have a column with the local time, but only UTC time - then change lines 42-44 in Functions_RSS.Based.Localizations.R from 'Time.local' to 'Time' ************
##                    ***** BUT be sure that Dates and Times in TestInfo.csv are also in UTC time and not local time *********
##                -- NodeId: Factor identifying the node in the network that detected the specified tag
##                -- TagRSSI: Integer indicating the RSS value (in dB) of the signal of the specified tag received by the specified node
##
##       3. Nodes.csv: file with a list of all nodes in your network and their UTM locations. The file should be saved in the working directory defined below
##            - For this code, Diane's csv is: "Nodes_Example.csv"
##            - Columns
##                -- NodeId: unique identifier of given to each node 
#                 -- NodeUTMx: Easting location of the specified node
##                -- NodeUTMy: Northing location of the specified node 
##            - Other columns can also be in the file for your reference 
##            - If Node names are being converted to scientific notation in Excel open the file with a text editor (e.g. BBedit) to change names to the correct format and save the file  
## 
##       4. Functions_RSS.Based.Localizations.R - R script with functions needed to run the code below - file should be saved in the working directory defined below    
##      
##  Important Output
##      1. Calibration Dataset that contains the average RSS values for a given node that detected the signal of a test transmitter at a known location
##      2. Model coefficients from an exponential decay model that show the relationship between RSS and Distance for a node network
##            exponential model formula: avgRSS ~ a * exp(-S * distance) + K
##                  a = intercept
##                  S = decay factor
##                  K = horizontal asymptote
##
##  
##########################################################################################################################################################



# Packages needed
library(dplyr)
library(lubridate)
library(ggplot2)
library(stringr)

# Reset R's brain - removes all previous objects
rm(list=ls())


## Set by User
# Directory for Output data - Provide/path/where/you/want/output/data/to/be/stored/
outpath <- "data/localizations/"



## Bring in functions 
source("R_scripts/05_functions_rss.based.localizations.R")



## Bring in 3 Needed files - Test Information, RSS values, and Node Information - change file names in " " as needed

test.info <- read.csv("data-raw/2023_node_files/Test.Info_Example0.9.csv", header = T)
str(test.info) # check that data imported properly
test.info$TagId <- "6166071E" # had to change TagId value column to all one value because for some reason R was importing it as 6166071 instead of 6166071E
str(test.info) # check that data imported properly
# Subsetting test.info based on whether it is the calibration (A) or test dataset (B)
#test.info <- test.info[test.info$TestAB == 'A', ]
test.info <- test.info[test.info$PropTest == 'A', ]

beep.dat <- readRDS("data/beeps/BeepMerge.rds") 
str(beep.dat) # check that data imported properly

nodes <- read.csv("data-raw/2023_node_files/Nodes_Example.csv", header = T)
str(nodes) # check that data imported properly


#################################################################
#   Step 1
# Run function to isolate raw RSS data from a node network that
# is associated with test data  
#################################################################

## Variables to define for function
## TEST.TYPE = category indicating the type of dataset
## "Calibration" - test dataset used for calibrating the relationship between RSS and Distance (purpose of this R script)
## "LocError" - test data used to determine localization error associated with RSS-based localization estimates for a node network
## DATE.FORMAT = format of the date column in Test.Info.csv file using R standard date expressions (e.g., "%m-%d-%y", "%Y-%m-%d")
## TIMEZONE = Time zone where data was collected, use grep("<insert location name here>", OlsonNames(), value=TRUE) to find valid time zone name


## Output in R environment
## combined.data - dataframe that contains the average RSS values for a given node associated with each unique test
## Columns:
## NodeId - unique identifier of the specified node
## TestId - unique identifier of the specified test
## avgRSS - average RSS value across a 3-min time period for the given node and test
## sdRSS - standard deviation of RSS values across the 3-min time period for the given node and test
## distance - true distance (in meters) between the specified node and test location
## NodeUTMx - Easting location of the specified node
## NodeUTMy - Northing location of the specified node
## TestUTMx - Easting location of the specified test
## TestUTMy - Northing location of the specified test

## Output saved
## Calibration_Dataset.csv - .csv file of the dataframe 'combine.data' saved in the folder specified by the outpath


##******* Define Variables - replace values below with user specified values *******## 
TEST.TYPE <- "Calibration"
DATE.FORMAT <- "%m/%d/%Y" # changed Paxton code to have year represented as %Y due to putting in year as 2023 rather than 23 (which would require %y)
TIME.ZONE <- "America/New_York" 


# Combine RSS data from a node network with test information into a dataframe
combined.data <- data.setup(TEST.TYPE, DATE.FORMAT, TIME.ZONE)
str(combined.data)




##########################################################
#    Step 2
# Exponential Decay Function to Examine Relationship 
# between distance and Tag RSS values 
##########################################################


## Visualize data

# Plot of the relationship between RSS and distance
ggplot(data = combined.data, aes(x = distance, y = avgRSS, color = NodeId)) +
  geom_point(size = 2)


## Preliminary Exponential Decay Model to determine starting values for the final model 

# SSasymp - self start for exponential model to finding the starting values of the data
# Asym - horizontal asymptote (when large values ) - y values decay to this value 
# R0 - numeric value when avgRSS (i.e., response variable) = 0 
# lrc - natural logarithm of the rate constant (rate of decay)


# Preliminary Model
exp.mod <- nls(avgRSS ~ SSasymp(distance, Asym, R0, lrc), data = combined.data)
# Summary of Model
summary(exp.mod)
# rate of decay
exp(coef(exp.mod)[["lrc"]]) 
#0.005318208
#0.005617166 : AB
#0.005320165 :0.8 As
#0.00528154: 0.85 As
#0.005304291: 0.9 As

#Diane's 1st Output
#Parameters:
#       Estimate Std. Error t value Pr(>|t|)    
#Asym -104.34460    0.25045  -416.6   <2e-16 ***
#R0    -54.38621    0.73690   -73.8   <2e-16 ***
#lrc    -5.23662    0.02449  -213.9   <2e-16 ***
#  ---
# Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#
# Residual standard error: 7.073 on 2807 degrees of freedom
#
#Number of iterations to convergence: 7 
#Achieved convergence tolerance: 3.749e-06

#Diane's parameters after using just As from combined.data (AB)
##Parameters:
#   Estimate Std. Error t value Pr(>|t|)    
#Asym -103.91927    0.29250 -355.28   <2e-16 ***
# R0    -52.60159    0.92728  -56.73   <2e-16 ***
# lrc    -5.18193    0.02884 -179.68   <2e-16 ***
#  ---
#  Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

#Residual standard error: 7.179 on 1994 degrees of freedom

#Number of iterations to convergence: 8 
#Achieved convergence tolerance: 3.114e-06

#Diane's parameters for 0.8 As
#Parameters:
#      Estimate Std. Error t value Pr(>|t|)    
# Asym -104.28136    0.27905 -373.71   <2e-16 ***
#  R0    -54.28493    0.82147  -66.08   <2e-16 ***
# lrc    -5.23625    0.02724 -192.23   <2e-16

#Diane's parameters for 0.85 As
#Parameters:
#  Estimate Std. Error t value Pr(>|t|)    
#Asym -104.2844     0.2705 -385.57   <2e-16 ***
#  R0    -54.6306     0.7943  -68.78   <2e-16 ***
#  lrc    -5.2435     0.0266 -197.09   <2e-16 ***
#  ---
#  Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

#Residual standard error: 7.001 on 2403 degrees of freedom

#Number of iterations to convergence: 7 
#Achieved convergence tolerance: 3.206e-06

#Diane's parameters for 0.9 As
#Parameters:
#       Estimate Std. Error t value Pr(>|t|)    
# Asym -104.23413    0.26497 -393.38   <2e-16 ***
# R0    -54.42834    0.77369  -70.35   <2e-16 ***
# lrc    -5.23924    0.02591 -202.25   <2e-16 ***
#  ---
#  Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

#Residual standard error: 7.05 on 2513 degrees of freedom

#Number of iterations to convergence: 7 
#Achieved convergence tolerance: 3.251e-06



## Final Exponential Decay Model with user provided self-starting values 
## based on visualization of the data and values in the Preliminary Model Output

# exponential model formula: avgRSS ~ a * exp(-S * distance) + K
# a = intercept #Diane interpreted as the 'R0' above
# S = decay factor  #Diane interpreted as the 'rate of decay' calculated above
# K = horizontal asymptote #Diane interpreted as the 'Asym' above


##  ***** Variables to define for final model below - replace values below with values from exp.mod ****  ## 
a <- -54.42834 # -54.38621 for the regular run
S <- 0.005304291 #0.005318208
K <- -104.23413 #-104.34460


# Final Model
nls.mod <- nls(avgRSS ~ a * exp(-S * distance) + K, start = list(a = a, S = S, K= K), 
               data = combined.data)
# Model Summary
summary(nls.mod)
# Model Coefficients 
# **** you will use these coefficient values to estimate distance in Github_TestDataset_LocalizationError.R and Github_Simulations.R scripts *****
coef(nls.mod)
#            a             S             K 
# 4.995845e+01  5.318227e-03 -1.043446e+02 

#AB run
#a             S             K 
#5.131773e+01  5.617182e-03 -1.039193e+02 

#80% proportion run
#            a             S             K 
#4.999648e+01  5.320181e-03 -1.042813e+02 

#85% proportion run
#a             S             K 
#4.965384e+01  5.281557e-03 -1.042844e+02 

#90% proportion run
#a             S             K 
#4.980584e+01  5.304308e-03 -1.042341e+02 

## Check the fit of the model and get predicted values

# Get residuals and fit of model and add variables to main table
combined.data$E <- residuals(nls.mod)
combined.data$fit <- fitted(nls.mod)

# Plot residuals by fit or distance
ggplot(combined.data, aes(x = distance, y = E, color = NodeId)) +
  geom_point(size = 2)
ggplot(combined.data, aes(x = fit, y = E, color = NodeId)) +
  geom_point(size = 2)

# Get model predictions
combined.data$pred <- predict(nls.mod)


## Save Final Dataset with Residuals and Predictions
write.csv(combined.data, paste0(outpath, "Calibration_Dataset_withResiduals_Predictions90p.csv"),
          row.names = F)


## Plot with predicted line

pdf(paste0(outpath, "Relationship_Distance~RSSI.90p.pdf"), width = 8, height = 6)

ggplot(combined.data, aes(x = distance, y = avgRSS)) + 
  geom_point() +
  geom_line(aes(y = pred), color = "#689FBB", lwd = 1.25) +
  scale_y_continuous(name = "RSS (dB)") +
  scale_x_continuous(name = "Distance (m)") +
  theme_classic()

dev.off()
