==> R_scripts/15_rsf_analysis.R <==
###################################################################################
## Diane Klement
## May 10 2024
##
## Code to analyze the data generated by "14_rsf_spatial.R" script as a GLMM
##
###################################################################################

# Loading necessary packages
library(tidyverse)
library(lme4)
library(magrittr)
library(ggeffects)
library(sjmisc)
library(splines)
library(coxme)
library(car) 
library(survival)
library(broom.mixed)
library(rlang)
library(kableExtra)

# Reset R's brain - removes all previous objects
rm(list=ls())

################################################################################
# Reading in data
d <- readRDS("data/rsf/rsf_data_burn.rds")

################################################################################

# Re-editing data
d$veg_reclass_int <- d$veg_reclass
d$veg_reclass_int1 <- floor(d$veg_reclass_int)
str(d)

# 'veg_reclass' is stored as numeric but should be factors
d$veg_reclass_int1 <- as.factor(d$veg_reclass_int1)
class(d$veg_reclass_int)
levels(d$veg_reclass_int)
str(d)

# burn_reclass storing as a factor
d$burn_reclass <- as.factor(d$burn_reclass)
levels(d$burn_reclass)
d$burn_reclass <- factor(d$burn_reclass, levels=c('7', '3', '6', '2', '4'))
str(d)

d$year_two_three <- d$Year2_3
d$year_zero_one <- d$Year0_1
d$year_three_four <- d$Year3_4
d$year_one_two <- d$Year1_2

str(d)

d1 <- d %>%
        select(TagId, Used, burn_reclass)

tag <- readRDS("data/tagged_birds.rds")
tags <- tag[c("AUX_TAG_NO", "AGE", "SEX")] 
tag1 <- tags %>%
          rename(TagId = AUX_TAG_NO)

d_combined <- left_join (d, tag1, by = "TagId")
str(d_combined)
d_combined$AGE <- as.factor(d_combined$AGE)
d_combined$SEX <- as.factor(d_combined$SEX)
################################################################################
# GLM with sex and age too 

burn.mod <- glmer(Used ~ Year0_1 + Year1_2 + Year2_3 + Year3_4 + (1|TagId), data = d_combined, family = 'binomial')
burn.mod.m <- glm(Used ~ Year0_1 + Year1_2 + Year2_3 + Year3_4, data = d_combined, family = binomial(link="logit"))
d_combined$burn_reclass <- as.factor(d_combined$burn_reclass)
burn.model <- glm(Used ~ burn_reclass + AGE + SEX, data = d_combined, family = binomial(link="logit"))
summary(burn.model)
summary(burn.mod)
summary(burn.mod.m)
stats.table <- summary(burn.mod)

################################################################################
# Creating a GLMM

# Running univariate models

## USE THIS ONE FOR REPORTING AND TABLE
# New models for burn
#model with burn type
mBurn <- glmer(Used ~ burn_reclass + (1|TagId), data = d, family = 'binomial')
stats.table <- summary(mBurn)
#               Estimate Std. Error z value Pr(>|z|)  
#(Intercept)   -0.07563    0.12359  -0.612   0.5406     unburned
#burn_reclass3  0.09058    0.15473   0.585   0.5583     year0-1
#burn_reclass6 -0.32529    0.15648  -2.079   0.0376 *   year1-2
#burn_reclass2  0.08061    0.11868   0.679   0.4970     year2-3
#burn_reclass4  0.59560    0.74498   0.799   0.4240     year3-4

# pr(use) of  1-2 yr grassland
plogis(-0.07563 -0.32529) #0.4010913
plogis(-0.07563) #0.4811015
lowerval <- plogis(-0.07563 -0.32529 - 1.96 * 0.15648) #0.3301248
upperval <- plogis(-0.07563 -0.32529 + 1.96 * 0.15648) #0.4764626
exp(-0.32529)

nrow(d)
used <- d %>%
          filter(burn_reclass == "2")
nrow(used)

df<- data.frame(Location = c("Bass Creek Grid", "Bass Creek Grid", "Bass Creek Grid", 
                             "Mosquito Creek Grid", "Mosquito Creek Grid", "Mosquito Creek Grid"),
                Age_Sex = c("After-second-year (ASY) male", "After-second-year (ASY) female", "Second-year (SY) female",
                            "After-second-year (ASY) male", "After-second-year (ASY) female", "Second-year (SY) female"),
                Tags = c(4, 2, 3, 1, 6, 2),
                Total = c("22.22%", "11.11%", "16.67%", 
                          "5.56%", "33.33%", "11.11%")
                )
stats.table <- kableExtra::kable(df, conf.int = FALSE)
tab <- knitr::kable(stats.table)
my_table <- rempsyc::nice_table(df,
                                title = c("Tagged Painted Bunting distribution"),
                                note = c("* p < .05, ** p < .01"))
print(my_table, preview = "docx")

mBurn.glm <- glm(Used ~ burn_reclass, family = binomial(link = "logit"), data = d1)
summary(mBurn.glm)
#Coefficients:
#               Estimate Std. Error z value Pr(>|z|)  
#(Intercept)   -0.04134    0.07976  -0.518   0.6042     unburned
#burn_reclass3  0.12247    0.10428   1.174   0.2402     year0-1
#burn_reclass6 -0.25863    0.11186  -2.312   0.0208 *   year1-2
#burn_reclass2  0.05368    0.09943   0.540   0.5893     year2-3
#burn_reclass4  0.55217    0.73452   0.752   0.4522     year3-4

model.matrix(mBurn)
# pr(use) of unburned grassland
plogis(-0.07563) #0.4811015

# pr(use) of grassland burned 0-1 years ago
plogis(-0.07563 + 0.09058) #0.5037374
# odds of use of 0-1 yr burn grassland relative to unburned
exp(0.09058) #1.094809

# pr(use) of grassland burned 1-2 years ago
plogis(-0.07563 - 0.32529) #0.4010913
# odds of use of 1-2 yr burn grassland relative to unburned
exp(-0.32529) #0.7223179

# pr(use) of grassland burned 2-3 years ago
plogis(-0.07563 + 0.08061) #0.501245
# odds of use of 2-3 yr burn grassland relative to unburned
exp(0.08061) #1.083948

# pr(use) of grassland burned 3-4 years ago
plogis(-0.07563 + 0.59560) #0.6271408
# odds of use of 3-4 yr burn grassland relative to unburned
exp(0.59560) #1.814119

# Model for burn type but written differently than mBurn just for context
#including all dummy codes except for the unburned (to use as the comparison factor)
mBurn2<- glmer(Used ~  Year0_1 + Year1_2 + Year2_3 + Year3_4
              + (1|TagId), data = d, family = 'binomial')
summary(mBurn2)
#             Estimate Std. Error z value Pr(>|z|)  
#(Intercept) -0.07563    0.12359  -0.612   0.5406     unburned
#Year0_1      0.09058    0.15473   0.585   0.5583  
#Year1_2     -0.32529    0.15648  -2.079   0.0376 *
#Year2_3      0.08061    0.11868   0.679   0.4970  
#Year3_4      0.59560    0.74498   0.799   0.4240 

# Model for burn type but without random effect for TagId
mBurn3<- glm(Used ~  Year0_1 + Year1_2 + Year2_3 + Year3_4
               , data = d, family = binomial(link="logit"))
summary(mBurn3)
#Coefficients:
#            Estimate Std. Error z value Pr(>|z|)  
#(Intercept) -0.04134    0.07976  -0.518   0.6042     unburned
#Year0_1      0.12247    0.10428   1.174   0.2402  
#Year1_2     -0.25863    0.11186  -2.312   0.0208 *
#Year2_3      0.05368    0.09943   0.540   0.5893  
#Year3_4      0.55217    0.73452   0.752   0.4522 

# Same as mBurn3 but was trying to use this to get around the predict() not working
mBurn4<- glm(Used ~  year_zero_one + year_one_two + year_two_three + year_three_four
             , data = d, family = binomial(link="logit"))
summary(mBurn4)
broom::tidy(mBurn4)

# Model for veg type
# Getting rid of any veg columns with NA
#d1 <- d[!is.na(d$veg_reclass_int1), ]
#model with veg type
#mVeg <- glmer(Used ~ veg_reclass_int1 + (1|TagId), data = d1, family = 'binomial')
#summary(mVeg)

# Model for veg type
#including all dummy codes except for most common one - Maritime Grassland
#mVegc<- glmer(Used ~  Fresh_Tidal_Marsh + Salt_Tidal_Marsh + Tidal_Wooded_Swamp + Mar_Forest + Ponds_Lakes + Interdune_Wetland + Beach + (1|TagId), data = d1, family = 'binomial')

#summary(mVegc)


################################################################################
# Predicting the data to try and make data visualizations
colors2 <- c("#E3E418FF", "#5DC863FF", "#2C728EFF","#3E4A89FF", "#481F70FF")

#An alternate attempt?
#Occurrence probability of time since burn
predData.burn.m <- data.frame(burn_reclass = c('2', '3', '4', '6', '7'))
predData.burn.m$burn_reclass <- as.factor(predData.burn.m$burn_reclass)
pred.burn.m <- stats::predict(mBurn.glm, newdata = predData.burn.m, se.fit = TRUE)
predData.burn.m$p <- plogis(pred.burn.m$fit) # back transform to the probability scale
predData.burn.m$lower <- plogis(pred.burn.m$fit - pred.burn.m$se.fit)
predData.burn.m$upper <- plogis(pred.burn.m$fit + pred.burn.m$se.fit)
predData.burn.m$burn_reclass <- factor(predData.burn.m$burn_reclass,
                                     levels = c('3', '6', '2', '4', '7'),
                                     labels = c("0-1 Years", "1-2 Years", "2-3 Years", "3-4 Years", "Unburned"))

predData.burn.f <- data.frame(burn_reclass = c('2', '3', '4', '6', '7'))
predData.burn.f$burn_reclass <- as.factor(predData.burn.f$burn_reclass)
pred.burn.f <- stats::predict(burn.model, newdata = predData.burn.f, se.fit = TRUE)
predData.burn.f$p <- plogis(pred.burn.f$fit) # back transform to the probability scale
predData.burn.f$lower <- plogis(pred.burn.f$fit - pred.burn.f$se.fit)
predData.burn.f$upper <- plogis(pred.burn.f$fit + pred.burn.f$se.fit)
predData.burn.f$burn_reclass <- factor(predData.burn.f$burn_reclass,
                                       levels = c('3', '6', '2', '4', '7'),
                                       labels = c("0-1 Years", "1-2 Years", "2-3 Years", "3-4 Years", "Unburned"))


tsb_node <-ggplot() +
  geom_col(data = predData.burn.m, aes(x = burn_reclass, y = p)) +
  geom_errorbar(data = predData.burn.m, aes(x = burn_reclass, ymin = lower, ymax = upper),
                width = 0.1) +
  scale_y_continuous("Probability of Use") +
  scale_x_discrete("Time Since Burn") +
  theme_classic()

tsb_nodef <-ggplot() +
  geom_col(data = predData.burn.f, aes(x = burn_reclass, y = p)) +
  geom_errorbar(data = predData.burn.f, aes(x = burn_reclass, ymin = lower, ymax = upper),
                width = 0.1) +
  scale_y_continuous("Probability of Use") +
  scale_x_discrete("Time Since Burn") +
  theme_classic()

# Saving plot
use_sexes <- ggpubr::ggarrange(tsb_node,
                                     tsb_nodef,
                                     nrow =1, ncol =2)

ggsave("data/figures/plots/tsb_node_gray.pdf", plot = tsb_node, width = 5, height = 5)
ggsave( "data/figures/plots/tsb_node_gray.jpeg", plot = tsb_node, width =5, height = 5, dpi = 600)
ggsave( "data/figures/plots/tsb_node_gray_female.jpeg", plot = tsb_nodef, width = 6, height = 6, dpi = 600)
